ENH: Use same system as in Builder to parse demos folder and create demos menu, meaning names are "prettified" and minimises duplicate code for each
